module5-task1 задание
