+++ index.html
@@ -235,7 +235,7 @@
     </section>
   </template>
   <script src="/js/functions.js"></script>
-  <script src="/js/main.js"></script>
+  <script type="module" src="/js/main.js"></script>
 
 </body>
